requirements.txt
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.1
networkx==3.1
pandas==2.0.3
matplotlib==3.8.0
seaborn==0.13.0
python-louvain==0.16
xgboost==2.0.2
pyyaml==6.0
tqdm==4.66.1
